chore(wasm-instrument): move syscall table test to core-backend (#3460)
