+++ utils/wasm-instrument/Cargo.toml
@@ -17,12 +17,10 @@ enum-iterator.workspace = true
 [dev-dependencies]
 wasmparser.workspace = true
 wat.workspace = true
-gear-core-backend = { workspace = true, features = ["mock"] }
 gear-core.workspace = true
 
 [features]
 default = ["std"]
 std = [
-    "gear-core-backend/std",
     "gwasm-instrument/std",
 ]
