+++ utils/wasm-instrument/src/tests.rs
@@ -618,61 +618,6 @@ test_gas_counter_injection! {
     "#
 }
 
-/// Check that all sys calls are supported by backend.
-#[test]
-fn test_sys_calls_table() {
-    use gas_metering::ConstantCostRules;
-    use gear_core::message::DispatchKind;
-    use gear_core_backend::{
-        env::{BackendReport, Environment},
-        error::ActorTerminationReason,
-        mock::MockExt,
-    };
-    use parity_wasm::builder;
-
-    // Make module with one empty function.
-    let mut module = builder::module()
-        .function()
-        .signature()
-        .build()
-        .build()
-        .build();
-
-    // Insert syscalls imports.
-    for name in SysCallName::instrumentable() {
-        let sign = name.signature();
-        let types = module.type_section_mut().unwrap().types_mut();
-        let type_no = types.len() as u32;
-        types.push(parity_wasm::elements::Type::Function(sign.func_type()));
-
-        module = builder::from_module(module)
-            .import()
-            .module("env")
-            .external()
-            .func(type_no)
-            .field(name.to_str())
-            .build()
-            .build();
-    }
-
-    let module = inject(module, &ConstantCostRules::default(), "env").unwrap();
-    let code = module.into_bytes().unwrap();
-
-    // Execute wasm and check success.
-    let ext = MockExt::default();
-    let env =
-        Environment::new(ext, &code, DispatchKind::Init, Default::default(), 0.into()).unwrap();
-    let report = env
-        .execute(|_, _, _| -> Result<(), u32> { Ok(()) })
-        .unwrap();
-
-    let BackendReport {
-        termination_reason, ..
-    } = report;
-
-    assert_eq!(termination_reason, ActorTerminationReason::Success.into());
-}
-
 #[test]
 fn check_memory_array_pointers_definition_correctness() {
     let sys_calls = SysCallName::instrumentable();
